Add OOP principles

+++ src/main/java/com/aegon/survey/demo/repository/SurveyRepository.java
@@ -1,12 +1,12 @@
 package com.aegon.survey.demo.repository;
 
+import com.aegon.survey.demo.entity.Survey;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import com.aegon.survey.demo.entity.Survey;
 
 @Repository
-public interface SurveyRepository extends JpaRepository<Survey, Integer> {
-	
-	Boolean existsByTopic(String topic);
+public interface SurveyRepository extends JpaRepository<Survey, Long> {
+
+    Boolean existsByTopic(String topic);
 
 }
